requirements.txt
fastapi
uvicorn
firebase-admin
pymongo
python-jose[cryptography]
pydantic
pydantic-settings
python-dotenv
requests
passlib[bcrypt]
